Functions work — still a lot of TODOS
The main functions — that is create, read, delete and modify — work so far. However there are still several issues.

First, the optional fields need to be taken into acount (especially in the create function).
Second: the names of the variables should be changed so that they are consistent with the OMAS lib (e.g. change firstname into givenname)
Third: it needs to be done that the „has_permissions“ can be also given in the json for the create function

further it has been adjustet that in the create function, the fields userid and inproject can be optional.
Also: the json form of a modify request is now the same form as the json in the create function (added the „has_permissions“)